requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
backoff==2.2.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
djangorestframework==3.15.2
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
googlemaps==4.10.0
idna==3.10
multidict==6.1.0
numpy==2.2.1
opencage==3.0.4
pandas==2.2.3
propcache==0.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
yarl==1.18.3
